+++ CHANGELOG.md
@@ -39,7 +39,8 @@
     + Add `es_` prefix for NSTimer methods: `-es_timerWithTimeInterval:`, `-es_scheduledTimerWithTimeInterval:`
 - Removed:
     + `ESLocalizedString()`, `_e()`, `ESLocalizedStringWithFormat()`
-    + `ESGetAssociatedObject()`, `ESSetAssociatedObject()`, `OBJC_ASSOCIATION_WEAK`, use `objc_getAssociatedObject`, `objc_setAssociatedObject` directly
+    + `ESGetAssociatedObject()`, `ESSetAssociatedObject()`, use `objc_getAssociatedObject`, `objc_setAssociatedObject` directly
+    + `ESDefineAssociatedObjectKey`, `OBJC_ASSOCIATION_WEAK`
     + `ESOSVersionIsAbove*()`, `ESStringFromSize()`, `NSStringWith()`, `UIImageFromCache()`, `UIImageFrom()`
     + `NSObject (ESAssociatedObjectHelper)` category
     + `NSRegularExpression (ESAdditions)` category